requirements.txt
ruamel.yaml
iso3166
jsonschema
pytest
